fix: Make Cap.from_name() checks stricter
Now it outright won't accept names if the portion after "CAP_" does not
start with a capital letter.

The main concern here is hidden attribute access since the name is fed
to getattr(). That probably isn't an issue because the name is converted
to uppercase, so it would only work for attributes with uppercase names
anyway, but let's be careful.